+++ src/mento.test.ts
@@ -164,12 +164,12 @@ describe('Mento', () => {
 
   describe('createWithBrokerAddress', () => {
     it('should return a Mento instance without querying the registry', () => {
-      const testee = Mento.createWithBrokerAddress(fakeBrokerAddr, provider)
+      const testee = Mento.createWithParams(provider, fakeBrokerAddr)
       expect(testee).toBeDefined()
       expect(mockContractModule).toHaveBeenCalledTimes(0)
       expect(fakeRegistryContract.getAddressForString).toHaveBeenCalledTimes(0)
 
-      const testee2 = Mento.createWithBrokerAddress(fakeBrokerAddr, signer)
+      const testee2 = Mento.createWithParams(signer, fakeBrokerAddr)
       expect(testee2).toBeDefined()
       expect(mockContractModule).toHaveBeenCalledTimes(0)
       expect(fakeRegistryContract.getAddressForString).toHaveBeenCalledTimes(0)
@@ -177,13 +177,13 @@ describe('Mento', () => {
 
     it('should throw if the signer has no provider', () => {
       expect(() =>
-        Mento.createWithBrokerAddress(fakeBrokerAddr, signerWithoutProvider)
+        Mento.createWithParams(signerWithoutProvider, fakeBrokerAddr)
       ).toThrow('Signer must be connected to a provider')
     })
 
     it('should throw if no signer or provider is provided', () => {
       //@ts-ignore
-      expect(() => Mento.createWithBrokerAddress(fakeBrokerAddr)).toThrow(
+      expect(() => Mento.createWithParams(fakeBrokerAddr)).toThrow(
         'A valid signer or provider must be provided'
       )
     })
